+++ modules/ingester/instance.go
@@ -16,6 +16,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
+	"go.uber.org/atomic"
 	"google.golang.org/grpc/codes"
 
 	"github.com/grafana/tempo/modules/overrides"
@@ -53,8 +54,9 @@ var (
 )
 
 type instance struct {
-	tracesMtx sync.Mutex
-	traces    map[uint32]*trace
+	tracesMtx  sync.Mutex
+	traces     map[uint32]*trace
+	traceCount atomic.Int32
 
 	blocksMtx        sync.RWMutex
 	headBlock        *wal.AppendBlock
@@ -94,6 +96,12 @@ func newInstance(instanceID string, limiter *Limiter, writer tempodb.Writer, l *
 }
 
 func (i *instance) Push(ctx context.Context, req *tempopb.PushRequest) error {
+	// check for max traces before grabbing the lock to better load shed
+	err := i.limiter.AssertMaxTracesPerUser(i.instanceID, int(i.traceCount.Load()))
+	if err != nil {
+		return status.Errorf(codes.FailedPrecondition, "%s max live traces per tenant exceeded: %v", overrides.ErrorPrefixLiveTracesExceeded, err)
+	}
+
 	i.tracesMtx.Lock()
 	defer i.tracesMtx.Unlock()
 
@@ -107,14 +115,6 @@ func (i *instance) Push(ctx context.Context, req *tempopb.PushRequest) error {
 	}
 
 	return nil
-}
-
-// PushBytes is used by the wal replay code and so it can push directly into the head block with 0 shenanigans
-func (i *instance) PushBytes(ctx context.Context, id []byte, object []byte) error {
-	i.blocksMtx.Lock()
-	defer i.blocksMtx.Unlock()
-
-	return i.headBlock.Write(id, object)
 }
 
 // Moves any complete traces out of the map to complete traces
@@ -321,6 +321,16 @@ func (i *instance) FindTraceByID(id []byte) (*tempopb.Trace, error) {
 	return nil, nil
 }
 
+// AddCompletingBlock adds an AppendBlock directly to the slice of completing blocks.
+// This is used during wal replay. It is expected that calling code will add the appropriate
+// jobs to the queue to eventually flush these.
+func (i *instance) AddCompletingBlock(b *wal.AppendBlock) {
+	i.blocksMtx.Lock()
+	defer i.blocksMtx.Unlock()
+
+	i.completingBlocks = append(i.completingBlocks, b)
+}
+
 // getOrCreateTrace will return a new trace object for the given request
 //  It must be called under the i.tracesMtx lock
 func (i *instance) getOrCreateTrace(req *tempopb.PushRequest) (*trace, error) {
@@ -335,15 +345,11 @@ func (i *instance) getOrCreateTrace(req *tempopb.PushRequest) (*trace, error) {
 		return trace, nil
 	}
 
-	err = i.limiter.AssertMaxTracesPerUser(i.instanceID, len(i.traces))
-	if err != nil {
-		return nil, status.Errorf(codes.FailedPrecondition, "%s max live traces per tenant exceeded: %v", overrides.ErrorPrefixLiveTracesExceeded, err)
-	}
-
 	maxBytes := i.limiter.limits.MaxBytesPerTrace(i.instanceID)
 	trace = newTrace(maxBytes, fp, traceID)
 	i.traces[fp] = trace
 	i.tracesCreatedTotal.Inc()
+	i.traceCount.Inc()
 
 	return trace, nil
 }
@@ -376,6 +382,7 @@ func (i *instance) tracesToCut(cutoff time.Duration, immediate bool) []*trace {
 			delete(i.traces, key)
 		}
 	}
+	i.traceCount.Store(int32(len(i.traces)))
 
 	return tracesToCut
 }